[GOData] Be strict on the number of dimensions.
The previous implementation allowed to handle a matrix as a vector,
or a vector as a matrix, but that was bogus, made the code more complex
and was probably not needed.

Now, when requesting a value at a given coordinate, the number of dimension
of the actual object has to match.