GOColorPalette: don't use widget background to store color.
This simplifies the code by using a gobject data item to store the color
instead of forcing that information into the gui.  The code is simpler
and less likely to break.

This might also fix a problem with history.  The activate handler will
now get the color from the same place, not the color the swatch had when
created.