Decimal64 support.
This is experimental.  There is no support available.  Please do not
file bug reports unless they have ready-to-commit patches.

This adds support for _Decimal64, a 16-digit base-10 floating-point
type.

goffice/math/go-decimal.64 includes half a libm for the type which
often goes via double to do its job.  Since _Decimal64 and double
have, roughly, the same precision that's not too bad.  However, it
would not be workable for the _Decimal128 type, so no support for
that.